deleted batchbuilder_internal_test.go
