+++ sequencer/sequencer_internal_test.go
@@ -1,302 +1,464 @@
 package sequencer
 
-// TODO: commented, until process batch is implemented
-//import (
-//	"context"
-//	"math/big"
-//	"os"
-//	"strings"
-//	"testing"
-//	"time"
-//
-//	"github.com/ethereum/go-ethereum/accounts/abi/bind"
-//	"github.com/ethereum/go-ethereum/common"
-//	"github.com/ethereum/go-ethereum/core/types"
-//	"github.com/ethereum/go-ethereum/crypto"
-//	"github.com/ethereum/go-ethereum/trie"
-//	cfgTypes "github.com/0xPolygonHermez/zkevm-node/config/types"
-//	"github.com/0xPolygonHermez/zkevm-node/db"
-//	"github.com/0xPolygonHermez/zkevm-node/encoding"
-//	ethmanTypes "github.com/0xPolygonHermez/zkevm-node/etherman/types"
-//	"github.com/0xPolygonHermez/zkevm-node/pool"
-//	"github.com/0xPolygonHermez/zkevm-node/pool/pgpoolstorage"
-//	"github.com/0xPolygonHermez/zkevm-node/pricegetter"
-//	"github.com/0xPolygonHermez/zkevm-node/sequencer/profitabilitychecker"
-//	"github.com/0xPolygonHermez/zkevm-node/state"
-//	"github.com/0xPolygonHermez/zkevm-node/state/tree"
-//	"github.com/0xPolygonHermez/zkevm-node/test/dbutils"
-//	"github.com/jackc/pgx/v4/pgxpool"
-//	"github.com/stretchr/testify/mock"
-//	"github.com/stretchr/testify/require"
-//)
-//
-//const (
-//	getPendingTxsLimit = 30
-//	maxTxsInSequence   = 5
-//)
-//
-//type stateTestInterface interface {
-//	stateInterface
-//	GetLastBatch(ctx context.Context, isVirtual bool, txBundleID string) (*state.Batch, error)
-//	GetNonce(ctx context.Context, address common.Address, batchNumber uint64, txBundleID string) (uint64, error)
-//	GetBalance(ctx context.Context, address common.Address, batchNumber uint64, txBundleID string) (*big.Int, error)
-//	NewGenesisBatchProcessor(genesisStateRoot []byte, txBundleID string) (*state.BatchProcessor, error)
-//	SetLastBatchNumberSeenOnEthereum(ctx context.Context, batchNumber uint64, txBundleID string) error
-//	SetGenesis(ctx context.Context, genesis state.Genesis, txBundleID string) error
-//}
-//
-//var (
-//	dbCfg     = dbutils.NewConfigFromEnv()
-//	stateDB   *pgxpool.Pool
-//	testState stateTestInterface
-//	seqCfg    Config
-//	pl        *pool.Pool
-//
-//	genesisHash common.Hash
-//	txs         []*types.Transaction
-//
-//	senderPrivateKey           = "0x28b2b0318721be8c8339199172cd7cc8f5e273800a35616ec893083a4b32c02e"
-//	addr                       = common.HexToAddress("b94f5374fce5edbc8e2a8697c15331677e6ebf0b")
-//	consolidatedTxHash         = common.HexToHash("0x125714bb4db48757007fff2671b37637bbfd6d47b3a4757ebbd0c5222984f905")
-//	maticCollateral            = big.NewInt(1000000000000000000)
-//	lastBatchNumberSeen uint64 = 1
-//)
-//
-//var stateCfg = state.Config{
-//	MaxCumulativeGasUsed: 800000,
-//}
-//
-//func setUpBlock(ctx context.Context, t *testing.T) {
-//	blockHash := common.HexToHash("0x65b4699dda5f7eb4519c730e6a48e73c90d2b1c8efcd6a6abdfd28c3b8e7d7d9")
-//
-//	block := &state.Block{
-//		BlockNumber: 1,
-//		BlockHash:   blockHash,
-//		ParentHash:  genesisHash,
-//		ReceivedAt:  time.Now(),
-//	}
-//
-//	_, err := stateDB.Exec(ctx, "INSERT INTO state.block (block_num, block_hash, parent_hash, received_at) VALUES ($1, $2, $3, $4)",
-//		block.BlockNumber, block.BlockHash.Bytes(), block.ParentHash.Bytes(), block.ReceivedAt)
-//	if err != nil {
-//		require.NoError(t, err)
-//	}
-//}
-//
-//func setUpBatch(ctx context.Context, t *testing.T) {
-//	receivedAt := time.Now().Add(time.Duration(-5) * time.Minute)
-//	consolidatedAt := time.Now()
-//	batch := &state.Batch{
-//		BlockNumber:        1,
-//		Sequencer:          addr,
-//		Aggregator:         addr,
-//		ConsolidatedTxHash: consolidatedTxHash,
-//		Header:             &types.Header{Number: big.NewInt(1)},
-//		Uncles:             nil,
-//		RawTxsData:         nil,
-//		MaticCollateral:    maticCollateral,
-//		ReceivedAt:         receivedAt,
-//		ConsolidatedAt:     &consolidatedAt,
-//		ChainID:            big.NewInt(1000),
-//		GlobalExitRoot:     common.Hash{},
-//	}
-//	bp, err := testState.NewGenesisBatchProcessor(nil, "")
-//	if err != nil {
-//		require.NoError(t, err)
-//	}
-//
-//	err = bp.ProcessBatch(ctx, batch)
-//	if err != nil {
-//		require.NoError(t, err)
-//	}
-//}
-//
-//func TestMain(m *testing.M) {
-//	var err error
-//
-//	if err := dbutils.InitOrReset(dbCfg); err != nil {
-//		panic(err)
-//	}
-//
-//	stateDB, err = db.NewSQLDB(dbCfg)
-//	if err != nil {
-//		panic(err)
-//	}
-//	defer stateDB.Close()
-//
-//	store := tree.NewPostgresStore(stateDB)
-//	mt := tree.NewMerkleTree(store, tree.DefaultMerkleTreeArity)
-//	scCodeStore := tree.NewPostgresSCCodeStore(stateDB)
-//	testState = state.NewState(stateCfg, state.NewPostgresStorage(stateDB), tree.NewStateTree(mt, scCodeStore))
-//
-//	seqCfg = Config{
-//		WaitPeriodPoolIsEmpty:              cfgTypes.NewDuration(time.Second),
-//		LastL1InteractionTimeMaxWaitPeriod: cfgTypes.NewDuration(60 * time.Second),
-//		WaitBlocksToUpdateGER:              10,
-//		MaxTimeForBatchToBeOpen:         cfgTypes.NewDuration(60 * time.Second),
-//		ProfitabilityChecker:               profitabilitychecker.Config{SendBatchesEvenWhenNotProfitable: true},
-//	}
-//
-//	s, err := pgpoolstorage.NewPostgresPoolStorage(dbCfg)
-//	if err != nil {
-//		panic(err)
-//	}
-//	pl = pool.NewPool(s, testState, stateCfg.L2GlobalExitRootManagerAddr)
-//
-//	genesisBlock := types.NewBlock(&types.Header{Number: big.NewInt(0)}, []*types.Transaction{}, []*types.Header{}, []*types.Receipt{}, &trie.StackTrie{})
-//	genesisBlock.ReceivedAt = time.Now()
-//	genesisHash = genesisBlock.Hash()
-//	balance, _ := big.NewInt(0).SetString("1000000000000000000000", encoding.Base10)
-//	genesis := state.Genesis{
-//		Block: genesisBlock,
-//		Balances: map[common.Address]*big.Int{
-//			common.HexToAddress("0x617b3a3528F9cDd6630fd3301B9c8911F7Bf063D"): balance,
-//		},
-//	}
-//
-//	err = testState.SetLastBatchNumberSeenOnEthereum(context.Background(), lastBatchNumberSeen, "")
-//	if err != nil {
-//		panic(err)
-//	}
-//	ctx := context.Background()
-//	_, err = stateDB.Exec(ctx, "DELETE FROM state.block")
-//	if err != nil {
-//		panic(err)
-//	}
-//
-//	err = testState.SetGenesis(ctx, genesis, "")
-//	if err != nil {
-//		panic(err)
-//	}
-//	privateKey, err := crypto.HexToECDSA(strings.TrimPrefix(senderPrivateKey, "0x"))
-//	if err != nil {
-//		panic(err)
-//	}
-//
-//	auth, err := bind.NewKeyedTransactorWithChainID(privateKey, big.NewInt(1000))
-//	if err != nil {
-//		panic(err)
-//	}
-//
-//	for i := 0; i < 10; i++ {
-//		tx := types.NewTransaction(uint64(i), common.Address{}, big.NewInt(10), uint64(21000), big.NewInt(10), []byte{})
-//		signedTx, err := auth.Signer(auth.From, tx)
-//		if err != nil {
-//			panic(err)
-//		}
-//		ctx := context.Background()
-//		if err := pl.AddTx(ctx, *signedTx); err != nil {
-//			panic(err)
-//		}
-//		txs = append(txs, signedTx)
-//	}
-//
-//	result := m.Run()
-//	os.Exit(result)
-//}
-//
-//func TestSequencerBaseFlow(t *testing.T) {
-//	// set up prerequisites
-//	txManager := new(txmanagerMock)
-//	ctx := context.Background()
-//	setUpBlock(ctx, t)
-//	setUpBatch(ctx, t)
-//	ethMan := new(ethermanMock)
-//	gasCostMin := big.NewInt(0)
-//	gasCostMax, _ := new(big.Int).SetString("1000000000000000000", 10)
-//	ethMan.On("EstimateGasSequenceBatches", mock.Anything).Return(gasCostMax, nil)
-//	ethMan.On("GetSendSequenceFee").Return(gasCostMin, nil)
-//
-//	// get pending txs to build sequences to send, so test could check if variables are equal
-//	pendTxs, err := pl.GetPendingTxs(ctx, false, getPendingTxsLimit)
-//	require.NoError(t, err)
-//	sequencesToSent := make([]ethmanTypes.Sequence, 5)
-//	for i := 0; i < 2; i++ {
-//		for k := maxTxsInSequence * i; k < maxTxsInSequence*(i+1); k++ {
-//			sequencesToSent[i].Txs = append(sequencesToSent[i].Txs, pendTxs[k].Transaction)
-//		}
-//	}
-//
-//	txManager.On("SequenceBatches", mock.MatchedBy(func(sequences []ethmanTypes.Sequence) bool {
-//		res := true
-//		for i := 0; i < len(sequences); i++ {
-//			for k := 0; k < len(sequences[i].Txs); k++ {
-//				res = res && sequences[i].Txs[k].Hash() == sequencesToSent[i].Txs[k].Hash()
-//			}
-//		}
-//		return res
-//	})).Return(nil, nil)
-//
-//	pg, err := pricegetter.NewClient(pricegetter.Config{
-//		Type:         pricegetter.DefaultType,
-//		DefaultPrice: pricegetter.TokenPrice{Float: new(big.Float).SetInt64(2000)},
-//	})
-//	require.NoError(t, err)
-//	seq, err := New(seqCfg, pl, testState, ethMan, pg, txManager)
-//	require.NoError(t, err)
-//	require.Equal(t, 10, len(pendTxs))
-//
-//	// try to process transactions that should fit in one sequence
-//	ticker := time.NewTicker(seqCfg.WaitPeriodPoolIsEmpty.Duration)
-//	for i := 0; i < maxTxsInSequence; i++ {
-//		seq.tryToProcessTx(ctx, ticker)
-//	}
-//
-//	// check, that all requested txs are selected and there are 5 txs in sequence in progress
-//	pendTxs, err = pl.GetPendingTxs(ctx, false, getPendingTxsLimit)
-//	require.NoError(t, err)
-//	require.Equal(t, 5, len(pendTxs))
-//	require.Equal(t, maxTxsInSequence, len(seq.sequenceInProgress.Txs))
-//	require.Equal(t, 0, len(seq.closedSequences))
-//
-//	// checks that if seq meets WaitBlocksToUpdateGER condition, it will close a sequence
-//	seq.cfg.MaxTimeForBatchToBeOpen = cfgTypes.NewDuration(0)
-//	seq.tryToProcessTx(ctx, ticker)
-//	pendTxs, err = pl.GetPendingTxs(ctx, false, getPendingTxsLimit)
-//	require.NoError(t, err)
-//
-//	// check, that after processing one tx, only 4 pending txs left in the pool,
-//	// seq have 1 tx in sequence in progress and 1 closed sequence
-//	require.Equal(t, 4, len(pendTxs))
-//	require.Equal(t, 1, len(seq.sequenceInProgress.Txs))
-//	require.Equal(t, 1, len(seq.closedSequences))
-//	require.Equal(t, maxTxsInSequence, len(seq.closedSequences[0].Txs))
-//
-//	// return config param back
-//	seq.cfg.MaxTimeForBatchToBeOpen = cfgTypes.NewDuration(10 * time.Second)
-//
-//	for i := 0; i < 4; i++ {
-//		seq.tryToProcessTx(ctx, ticker)
-//	}
-//
-//	// set config params that way, that txs will be selected and sent to ethereum
-//	seq.cfg.LastL1InteractionTimeMaxWaitPeriod = cfgTypes.NewDuration(0)
-//	seq.cfg.MaxTimeForBatchToBeOpen = cfgTypes.NewDuration(0)
-//	seq.tryToProcessTx(ctx, ticker)
-//
-//	// checks, that after processing there is no pending txs left, no txs in sequence in progress
-//	// and no pending closed sequences
-//	pendTxs, err = pl.GetPendingTxs(ctx, false, getPendingTxsLimit)
-//	require.NoError(t, err)
-//	require.Equal(t, 0, len(pendTxs))
-//	require.Equal(t, 0, len(seq.sequenceInProgress.Txs))
-//	require.Equal(t, 0, len(seq.closedSequences))
-//
-//	txManager.AssertNumberOfCalls(t, "SequenceBatches", 1)
-//	cleanUpBatches(ctx, t)
-//	cleanUpBlocks(ctx, t)
-//}
-//
-//func cleanUpBatches(ctx context.Context, t *testing.T) {
-//	_, err := stateDB.Exec(ctx, "DELETE FROM state.batch WHERE block_num = $1", 1)
-//	if err != nil {
-//		require.NoError(t, err)
-//	}
-//}
-//
-//func cleanUpBlocks(ctx context.Context, t *testing.T) {
-//	_, err := stateDB.Exec(ctx, "DELETE FROM state.block WHERE block_num = $1", 1)
-//	if err != nil {
-//		require.NoError(t, err)
-//	}
-//}
+import (
+	"context"
+	"math/big"
+	"testing"
+	"time"
+
+	cfgTypes "github.com/0xPolygonHermez/zkevm-node/config/types"
+	ethManTypes "github.com/0xPolygonHermez/zkevm-node/etherman/types"
+	"github.com/0xPolygonHermez/zkevm-node/pool"
+	sequencerMocks "github.com/0xPolygonHermez/zkevm-node/sequencer/mocks"
+	"github.com/0xPolygonHermez/zkevm-node/sequencer/profitabilitychecker"
+	"github.com/0xPolygonHermez/zkevm-node/state"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/stretchr/testify/require"
+)
+
+func TestIsSynced(t *testing.T) {
+	st := new(sequencerMocks.StateMock)
+	eth := new(sequencerMocks.EthermanMock)
+	s := Sequencer{state: st, etherman: eth}
+	ctx := context.Background()
+	st.On("GetLastVirtualBatchNum", ctx, nil).Return(uint64(1), nil)
+	eth.On("GetLatestBatchNumber").Return(uint64(1), nil)
+	isSynced := s.isSynced(ctx)
+	require.Equal(t, true, isSynced)
+	st.AssertExpectations(t)
+	eth.AssertExpectations(t)
+}
+
+func TestIsNotSynced(t *testing.T) {
+	st := new(sequencerMocks.StateMock)
+	eth := new(sequencerMocks.EthermanMock)
+	s := Sequencer{state: st, etherman: eth}
+	ctx := context.Background()
+	st.On("GetLastVirtualBatchNum", ctx, nil).Return(uint64(1), nil)
+	eth.On("GetLatestBatchNumber").Return(uint64(2), nil)
+	isSynced := s.isSynced(ctx)
+	require.Equal(t, false, isSynced)
+	st.AssertExpectations(t)
+	eth.AssertExpectations(t)
+}
+
+func TestShouldCloseSequenceTooBig(t *testing.T) {
+	s := Sequencer{}
+	s.sequenceInProgress = ethManTypes.Sequence{IsSequenceTooBig: true}
+	ctx := context.Background()
+	shouldClose := s.shouldCloseSequenceInProgress(ctx)
+	require.False(t, s.sequenceInProgress.IsSequenceTooBig)
+	require.True(t, shouldClose)
+}
+
+func TestShouldCloseSequenceReachedMaxAmountOfTxs(t *testing.T) {
+	s := Sequencer{}
+	txs := make([]types.Transaction, 0, maxTxsPerBatch)
+	for i := uint64(0); i < maxTxsPerBatch; i++ {
+		tx := types.NewTransaction(i, common.Address{}, big.NewInt(10), uint64(1), big.NewInt(10), []byte{})
+		txs = append(txs, *tx)
+	}
+	s.sequenceInProgress = ethManTypes.Sequence{Txs: txs}
+	ctx := context.Background()
+	shouldClose := s.shouldCloseSequenceInProgress(ctx)
+	require.True(t, shouldClose)
+}
+
+func TestShouldCloseDueToNewDepositsUpdateGER(t *testing.T) {
+	st := new(sequencerMocks.StateMock)
+	eth := new(sequencerMocks.EthermanMock)
+	dbTx := new(sequencerMocks.DbTxMock)
+	s := Sequencer{cfg: Config{WaitBlocksToUpdateGER: 10}, state: st, etherman: eth}
+	ctx := context.Background()
+	mainnetExitRoot := common.HexToHash("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a53cf2d7d9f1")
+	lastGer := &state.GlobalExitRoot{
+		BlockNumber:       1,
+		GlobalExitRootNum: big.NewInt(2),
+		MainnetExitRoot:   common.HexToHash("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9f1"),
+		RollupExitRoot:    common.HexToHash("0x30a885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9a0"),
+		GlobalExitRoot:    common.HexToHash("0x40a885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9a0"),
+	}
+	s.sequenceInProgress.GlobalExitRoot = lastGer.GlobalExitRoot
+	st.On("GetBlockNumAndMainnetExitRootByGER", ctx, s.sequenceInProgress.GlobalExitRoot, nil).Return(lastGer.BlockNumber, mainnetExitRoot, nil)
+	st.On("GetLatestGlobalExitRoot", ctx, nil).Return(lastGer, nil)
+	eth.On("GetLatestBlockNumber", ctx).Return(uint64(12), nil)
+	st.On("BeginStateTransaction", ctx).Return(dbTx, nil)
+	dbTx.On("Commit", ctx).Return(nil)
+	st.On("UpdateGERInOpenBatch", ctx, lastGer.GlobalExitRoot, dbTx).Return(nil).Once()
+	isShouldCloseDueToNewDeposits, err := s.shouldCloseDueToNewDeposits(ctx)
+	require.NoError(t, err)
+	require.Equal(t, false, isShouldCloseDueToNewDeposits)
+	st.AssertExpectations(t)
+	eth.AssertExpectations(t)
+	dbTx.AssertExpectations(t)
+}
+
+func TestShouldCloseDueToNewDepositsSequenceProfitable(t *testing.T) {
+	st := new(sequencerMocks.StateMock)
+	eth := new(sequencerMocks.EthermanMock)
+	profitabilityChecker := profitabilitychecker.New(profitabilitychecker.Config{SendBatchesEvenWhenNotProfitable: true}, nil, nil)
+	s := Sequencer{cfg: Config{WaitBlocksToUpdateGER: 10}, state: st, etherman: eth, checker: profitabilityChecker}
+	ctx := context.Background()
+	mainnetExitRoot := common.HexToHash("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a53cf2d7d9f1")
+	lastGer := &state.GlobalExitRoot{
+		BlockNumber:       1,
+		GlobalExitRootNum: big.NewInt(2),
+		MainnetExitRoot:   common.HexToHash("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9f1"),
+		RollupExitRoot:    common.HexToHash("0x30a885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9a0"),
+		GlobalExitRoot:    common.HexToHash("0x40a885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9a0"),
+	}
+	s.sequenceInProgress.GlobalExitRoot = lastGer.GlobalExitRoot
+	tx := types.NewTransaction(uint64(0), common.Address{}, big.NewInt(10), uint64(1), big.NewInt(10), []byte{})
+	s.sequenceInProgress.Txs = []types.Transaction{*tx}
+	st.On("GetBlockNumAndMainnetExitRootByGER", ctx, s.sequenceInProgress.GlobalExitRoot, nil).Return(lastGer.BlockNumber, mainnetExitRoot, nil)
+	st.On("GetLatestGlobalExitRoot", ctx, nil).Return(lastGer, nil)
+	eth.On("GetLatestBlockNumber", ctx).Return(uint64(12), nil)
+
+	isShouldCloseDueToNewDeposits, err := s.shouldCloseDueToNewDeposits(ctx)
+	require.NoError(t, err)
+	require.Equal(t, true, isShouldCloseDueToNewDeposits)
+	st.AssertExpectations(t)
+	eth.AssertExpectations(t)
+}
+
+func TestShouldCloseTooLongSinceLastVirtualized(t *testing.T) {
+	st := new(sequencerMocks.StateMock)
+	s := Sequencer{cfg: Config{MaxTimeForBatchToBeOpen: cfgTypes.NewDuration(1 * time.Second)}, state: st}
+	tx := types.NewTransaction(uint64(0), common.Address{}, big.NewInt(10), uint64(1), big.NewInt(10), []byte{})
+	s.sequenceInProgress.Txs = []types.Transaction{*tx}
+	s.sequenceInProgress.Timestamp = time.Now().Add(-s.cfg.MaxTimeForBatchToBeOpen.Duration).Unix()
+	ctx := context.Background()
+	lastBatchNumber := uint64(10)
+	st.On("GetLastBatchNumber", ctx, nil).Return(lastBatchNumber, nil)
+	st.On("IsBatchVirtualized", ctx, lastBatchNumber-1, nil).Return(true, nil)
+	isShouldCloseTooLongSinceLastVirtualized, err := s.shouldCloseTooLongSinceLastVirtualized(ctx)
+	require.NoError(t, err)
+	require.True(t, isShouldCloseTooLongSinceLastVirtualized)
+	st.AssertExpectations(t)
+}
+
+func TestCleanTxsIfTxsDataIsBiggerThanExpected(t *testing.T) {
+	s := &Sequencer{}
+	tx := types.NewTransaction(uint64(0), common.Address{}, big.NewInt(10), uint64(1), big.NewInt(10), []byte{})
+	for i := 0; i < 300; i++ {
+		s.sequenceInProgress.Txs = append(s.sequenceInProgress.Txs, *tx)
+	}
+	ctx := context.Background()
+	ticker := time.NewTicker(1 * time.Second)
+	err := s.cleanTxsIfTxsDataIsBiggerThanExpected(ctx, ticker)
+	require.NoError(t, err)
+	require.True(t, s.sequenceInProgress.IsSequenceTooBig)
+	// 1 transfer equals ~103 bytes, so 291 txs ~= 30000 bytes, what is maximum
+	require.Equal(t, 291, len(s.sequenceInProgress.Txs))
+}
+
+func TestCleanTxsIfTxsDataIsBiggerThanExpectedTxIsTooBig(t *testing.T) {
+	pl := new(sequencerMocks.PoolMock)
+	s := &Sequencer{pool: pl}
+	tx := types.NewTransaction(uint64(0), common.Address{}, big.NewInt(10), uint64(1), big.NewInt(10), []byte{})
+	txs := []types.Transaction{}
+	for i := 0; i < 300; i++ {
+		txs = append(txs, *tx)
+	}
+	data, err := state.EncodeTransactions(txs)
+	tx1 := types.NewTransaction(uint64(0), common.Address{}, big.NewInt(10), uint64(1), big.NewInt(10), data)
+	s.sequenceInProgress.Txs = []types.Transaction{*tx1}
+	require.NoError(t, err)
+	ctx := context.Background()
+	ticker := time.NewTicker(1 * time.Second)
+	pl.On("UpdateTxStatus", ctx, s.sequenceInProgress.Txs[0].Hash(), pool.TxStatusInvalid).Return(nil)
+	err = s.cleanTxsIfTxsDataIsBiggerThanExpected(ctx, ticker)
+	require.NoError(t, err)
+	require.False(t, s.sequenceInProgress.IsSequenceTooBig)
+	// 1 transfer equals ~103 bytes, so 291 txs ~= 30000 bytes, what is maximum
+	require.Equal(t, 0, len(s.sequenceInProgress.Txs))
+	pl.AssertExpectations(t)
+}
+
+func TestAppendPendingTxs(t *testing.T) {
+	pl := new(sequencerMocks.PoolMock)
+	ctx := context.Background()
+	s := &Sequencer{pool: pl}
+	minGasPrice := big.NewInt(1)
+	ticker := time.NewTicker(1 * time.Second)
+
+	poolTx := types.NewTransaction(uint64(1), common.Address{}, big.NewInt(10), uint64(1), big.NewInt(10), []byte{})
+	var poolTxs []*pool.Transaction
+	poolTxs = append(poolTxs, &pool.Transaction{Transaction: *poolTx})
+
+	pl.On("GetTxs", ctx, pool.TxStatusPending, false, minGasPrice.Uint64(), uint64(150)).Return(poolTxs, nil)
+	pendTxsAmount := s.appendPendingTxs(ctx, false, minGasPrice.Uint64(), 150, ticker)
+	require.Equal(t, 1, pendTxsAmount)
+	require.Equal(t, 1, len(s.sequenceInProgress.Txs))
+	pl.AssertExpectations(t)
+}
+
+func TestAppendPendingTxsFailedCounter(t *testing.T) {
+	pl := new(sequencerMocks.PoolMock)
+	ctx := context.Background()
+	s := &Sequencer{cfg: Config{MaxAllowedFailedCounter: 5}, pool: pl}
+	minGasPrice := big.NewInt(1)
+	ticker := time.NewTicker(1 * time.Second)
+
+	poolTx := types.NewTransaction(uint64(1), common.Address{}, big.NewInt(10), uint64(1), big.NewInt(10), []byte{})
+	var poolTxs []*pool.Transaction
+	poolTxs = append(poolTxs, &pool.Transaction{Transaction: *poolTx, FailedCounter: 55})
+
+	pl.On("GetTxs", ctx, pool.TxStatusPending, false, minGasPrice.Uint64(), uint64(150)).Return(poolTxs, nil)
+	pl.On("UpdateTxsStatus", ctx, []string{poolTxs[0].Hash().String()}, pool.TxStatusInvalid).Return(nil)
+	pendTxsAmount := s.appendPendingTxs(ctx, false, minGasPrice.Uint64(), 150, ticker)
+	require.Equal(t, uint64(0), pendTxsAmount)
+	require.Equal(t, 0, len(s.sequenceInProgress.Txs))
+	pl.AssertExpectations(t)
+}
+
+func TestProcessBatch(t *testing.T) {
+	st := new(sequencerMocks.StateMock)
+	s := &Sequencer{state: st}
+	dbTx := new(sequencerMocks.DbTxMock)
+	ctx := context.Background()
+	st.On("BeginStateTransaction", ctx).Return(dbTx, nil)
+	dbTx.On("Commit", ctx).Return(nil)
+	lastBatchNumber := uint64(10)
+	tx1 := *types.NewTransaction(0, common.Address{}, big.NewInt(10), uint64(1), big.NewInt(10), []byte{})
+	tx2 := *types.NewTransaction(1, common.HexToAddress("1"), big.NewInt(1), 0, big.NewInt(1), []byte("bbb"))
+
+	s.sequenceInProgress.Txs = []types.Transaction{tx1, tx2}
+
+	txsBatch1 := []*state.ProcessTransactionResponse{
+		{
+			TxHash:      tx1.Hash(),
+			Tx:          tx1,
+			IsProcessed: true,
+		},
+		{
+			TxHash:      tx2.Hash(),
+			Tx:          tx2,
+			IsProcessed: true,
+		},
+	}
+
+	processBatchResponse := &state.ProcessBatchResponse{
+		CumulativeGasUsed: 100000,
+		IsBatchProcessed:  true,
+		Responses:         txsBatch1,
+		NewStateRoot:      common.HexToHash("0x123"),
+		NewLocalExitRoot:  common.HexToHash("0x123"),
+	}
+	st.On("GetLastBatchNumber", ctx, dbTx).Return(lastBatchNumber, nil)
+	st.On("ProcessSequencerBatch", ctx, lastBatchNumber, s.sequenceInProgress.Txs, dbTx).Return(processBatchResponse, nil)
+	procResponse, err := s.processTxs(ctx)
+	require.NoError(t, err)
+	require.True(t, procResponse.isBatchProcessed)
+	require.Equal(t, 2, len(procResponse.processedTxs))
+	st.AssertExpectations(t)
+}
+
+func TestReprocessBatch(t *testing.T) {
+	st := new(sequencerMocks.StateMock)
+	s := &Sequencer{state: st}
+	dbTx := new(sequencerMocks.DbTxMock)
+	ctx := context.Background()
+	st.On("BeginStateTransaction", ctx).Return(dbTx, nil)
+	dbTx.On("Commit", ctx).Return(nil)
+	tx1 := *types.NewTransaction(0, common.Address{}, big.NewInt(10), uint64(1), big.NewInt(10), []byte{})
+	tx2 := *types.NewTransaction(1, common.HexToAddress("1"), big.NewInt(1), 0, big.NewInt(1), []byte("bbb"))
+	txs := []types.Transaction{tx1, tx2}
+	s.sequenceInProgress.Txs = txs
+
+	txsBatch1 := []*state.ProcessTransactionResponse{
+		{
+			TxHash:      txs[0].Hash(),
+			Tx:          txs[0],
+			IsProcessed: true,
+		},
+		{
+			TxHash:      txs[1].Hash(),
+			Tx:          txs[1],
+			IsProcessed: true,
+		},
+	}
+
+	processBatchResponse := &state.ProcessBatchResponse{
+		CumulativeGasUsed: 100000,
+		IsBatchProcessed:  true,
+		Responses:         txsBatch1,
+		NewStateRoot:      common.HexToHash("0x123"),
+		NewLocalExitRoot:  common.HexToHash("0x123"),
+	}
+
+	processedTxs, processedTxsHashes, unprocessedTxs, unprocessedTxsHashes := state.DetermineProcessedTransactions(processBatchResponse.Responses)
+
+	txsResponse := processTxResponse{
+		processedTxs:         processedTxs,
+		processedTxsHashes:   processedTxsHashes,
+		unprocessedTxs:       unprocessedTxs,
+		unprocessedTxsHashes: unprocessedTxsHashes,
+		isBatchProcessed:     false,
+	}
+	txsResponseToReturn := txsResponse
+	txsResponseToReturn.isBatchProcessed = true
+	lastBatchNumber := uint64(10)
+	st.On("GetLastBatchNumber", ctx, dbTx).Return(lastBatchNumber, nil)
+	st.On("ProcessSequencerBatch", ctx, lastBatchNumber, txs, dbTx).Return(processBatchResponse, nil)
+
+	unprocessedTxsAfterReprocess, err := s.reprocessBatch(ctx, txsResponse, ethManTypes.Sequence{})
+	require.NoError(t, err)
+	require.Equal(t, 0, len(unprocessedTxsAfterReprocess))
+	require.Equal(t, 2, len(s.sequenceInProgress.Txs))
+	st.AssertExpectations(t)
+}
+
+func TestUpdateTxsInPool(t *testing.T) {
+	st := new(sequencerMocks.StateMock)
+	pl := new(sequencerMocks.PoolMock)
+	s := &Sequencer{state: st, pool: pl}
+	ctx := context.Background()
+	ticker := time.NewTicker(1 * time.Second)
+
+	tx1 := *types.NewTransaction(0, common.Address{}, big.NewInt(10), uint64(1), big.NewInt(10), []byte{})
+	tx2 := *types.NewTransaction(1, common.HexToAddress("0x1"), big.NewInt(1), 0, big.NewInt(1), []byte("bbb"))
+	tx3 := *types.NewTransaction(3, common.HexToAddress("0x2"), big.NewInt(1), 0, big.NewInt(1), []byte("ddd"))
+
+	txs := []types.Transaction{tx1, tx2, tx3}
+	s.sequenceInProgress.Txs = txs
+
+	txsBatch1 := []*state.ProcessTransactionResponse{
+		{
+			TxHash:      txs[0].Hash(),
+			Tx:          txs[0],
+			IsProcessed: true,
+		},
+		{
+			TxHash:      txs[1].Hash(),
+			Tx:          txs[1],
+			IsProcessed: true,
+		},
+		{
+			TxHash:      txs[2].Hash(),
+			Tx:          txs[2],
+			IsProcessed: false,
+		},
+	}
+
+	processBatchResponse := &state.ProcessBatchResponse{
+		CumulativeGasUsed: 100000,
+		IsBatchProcessed:  true,
+		Responses:         txsBatch1,
+		NewStateRoot:      common.HexToHash("0x123"),
+		NewLocalExitRoot:  common.HexToHash("0x123"),
+	}
+
+	processedTxs, processedTxsHashes, unprocessedTxs, unprocessedTxsHashes := state.DetermineProcessedTransactions(processBatchResponse.Responses)
+
+	txsResponse := processTxResponse{
+		processedTxs:         processedTxs,
+		processedTxsHashes:   processedTxsHashes,
+		unprocessedTxs:       unprocessedTxs,
+		unprocessedTxsHashes: unprocessedTxsHashes,
+		isBatchProcessed:     false,
+	}
+
+	pl.On("UpdateTxsStatus", ctx, processedTxsHashes, pool.TxStatusSelected).Return(nil)
+
+	fromAddress := common.HexToAddress("0x123")
+	pl.On("GetTxFromAddressFromByHash", ctx, txs[2].Hash()).Return(fromAddress, txs[2].Nonce(), nil)
+	l2BlockNumber := uint64(3)
+	st.On("GetLastL2BlockNumber", ctx, nil).Return(l2BlockNumber, nil)
+	accNonce := uint64(2)
+	st.On("GetNonce", ctx, fromAddress, l2BlockNumber, nil).Return(accNonce, nil)
+	pl.On("UpdateTxsStatus", ctx, unprocessedTxsHashes, pool.TxStatusFailed).Return(nil)
+	pl.On("IncrementFailedCounter", ctx, unprocessedTxsHashes).Return(nil)
+	s.updateTxsInPool(ctx, ticker, txsResponse, unprocessedTxs)
+	st.AssertExpectations(t)
+	pl.AssertExpectations(t)
+}
+
+func TestTryToProcessTxs(t *testing.T) {
+	st := new(sequencerMocks.StateMock)
+	eth := new(sequencerMocks.EthermanMock)
+	dbTx := new(sequencerMocks.DbTxMock)
+	pl := new(sequencerMocks.PoolMock)
+
+	gpe := new(sequencerMocks.GasPriceEstimatorMock)
+	s := Sequencer{cfg: Config{MaxTimeForBatchToBeOpen: cfgTypes.NewDuration(5 * time.Second)}, state: st, etherman: eth, gpe: gpe, pool: pl}
+	ctx := context.Background()
+	// Check if synchronizer is up to date
+	st.On("GetLastVirtualBatchNum", ctx, nil).Return(uint64(1), nil)
+	eth.On("GetLatestBatchNumber").Return(uint64(1), nil)
+
+	isSynced := s.isSynced(ctx)
+	require.Equal(t, true, isSynced)
+
+	mainnetExitRoot := common.HexToHash("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a53cf2d7d9f1")
+	lastGer := &state.GlobalExitRoot{
+		BlockNumber:       1,
+		GlobalExitRootNum: big.NewInt(2),
+		MainnetExitRoot:   common.HexToHash("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a53cf2d7d9f1"),
+		RollupExitRoot:    common.HexToHash("0x30a885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9a0"),
+		GlobalExitRoot:    common.HexToHash("0x40a885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9a0"),
+	}
+	s.sequenceInProgress.GlobalExitRoot = lastGer.GlobalExitRoot
+	st.On("GetBlockNumAndMainnetExitRootByGER", ctx, s.sequenceInProgress.GlobalExitRoot, nil).Return(lastGer.BlockNumber, mainnetExitRoot, nil)
+	st.On("GetLatestGlobalExitRoot", ctx, nil).Return(lastGer, nil)
+	st.On("BeginStateTransaction", ctx).Return(dbTx, nil)
+	dbTx.On("Commit", ctx).Return(nil)
+
+	tx := types.NewTransaction(uint64(0), common.Address{}, big.NewInt(10), uint64(1), big.NewInt(10), []byte{})
+	s.sequenceInProgress.Txs = []types.Transaction{*tx}
+	s.sequenceInProgress.Timestamp = time.Now().Unix()
+
+	lastBatchNumber := uint64(10)
+	st.On("GetLastBatchNumber", ctx, nil).Return(lastBatchNumber, nil)
+	st.On("GetLastBatchNumber", ctx, dbTx).Return(lastBatchNumber, nil)
+
+	st.On("IsBatchVirtualized", ctx, lastBatchNumber-1, nil).Return(true, nil)
+
+	minGasPrice := big.NewInt(1)
+	gpe.On("GetAvgGasPrice", ctx).Return(minGasPrice, nil)
+
+	ticker := time.NewTicker(1 * time.Second)
+	poolTx := types.NewTransaction(uint64(1), common.Address{}, big.NewInt(10), uint64(1), big.NewInt(10), []byte{})
+	var poolTxs []*pool.Transaction
+	poolTxs = append(poolTxs, &pool.Transaction{Transaction: *poolTx})
+	pl.On("GetTxs", ctx, pool.TxStatusPending, true, uint64(0), uint64(149)).Return([]*pool.Transaction{}, nil)
+	pl.On("GetTxs", ctx, pool.TxStatusFailed, true, uint64(0), uint64(149)).Return([]*pool.Transaction{}, nil)
+
+	pl.On("GetTxs", ctx, pool.TxStatusPending, false, minGasPrice.Uint64(), uint64(149)).Return(poolTxs, nil)
+
+	txsBatch1 := []*state.ProcessTransactionResponse{
+		{
+			TxHash:      s.sequenceInProgress.Txs[0].Hash(),
+			Tx:          s.sequenceInProgress.Txs[0],
+			IsProcessed: true,
+		},
+		{
+			TxHash:      poolTxs[0].Transaction.Hash(),
+			Tx:          poolTxs[0].Transaction,
+			IsProcessed: true,
+		},
+	}
+
+	processBatchResponse := &state.ProcessBatchResponse{
+		CumulativeGasUsed: 100000,
+		IsBatchProcessed:  true,
+		Responses:         txsBatch1,
+		NewStateRoot:      common.HexToHash("0x123"),
+		NewLocalExitRoot:  common.HexToHash("0x123"),
+	}
+	var txs = s.sequenceInProgress.Txs
+	txs = append(txs, poolTxs[0].Transaction)
+	st.On("ProcessSequencerBatch", ctx, lastBatchNumber, txs, dbTx).Return(processBatchResponse, nil)
+
+	processedTxs, processedTxsHashes, _, _ := state.DetermineProcessedTransactions(processBatchResponse.Responses)
+
+	st.On("StoreTransactions", ctx, lastBatchNumber, processedTxs, dbTx).Return(nil)
+	pl.On("UpdateTxsStatus", ctx, processedTxsHashes, pool.TxStatusSelected).Return(nil)
+
+	s.tryToProcessTx(ctx, ticker)
+
+	st.AssertExpectations(t)
+	pl.AssertExpectations(t)
+	eth.AssertExpectations(t)
+}
